sky: add installable app manifest
this allows you to install this as a PWA on any modern device,
including laptops (via Chrome).

this also takes the document out of quirks mode
which required that i ditch +manx-response:server
for a builder that can add the necessary DOCTYPE directive.

current problem:

the start_url of the manifest is not able to be relative
when embedded in a data:... url or when served from s3.

to be correct, the json file (and like the hawk.png asset too)
must be served from the server itself. through a shrub.